Fix version field type consistency to u32
- Fixed remaining i64 to u32 conversions
- Updated memtable tuples to use Option<u32> for version
- Fixed SearchVectorRecord conversions
- Corrected version field usage across storage engines

+++ src/storage/engines/impls/viper/engine.rs
@@ -1098,7 +1098,7 @@ impl ViperEngine {
                             timestamp: timestamp as i64,
                             updated_at: Some(updated_at as i64),
                             expires_at: expires_at.map(|v| v as i64),
-                            version: Some(version as i64),
+                            version: Some(version as u32),
                             source: None,
                         };
                         // Check if this is a better match than what we have
